refactor(artifacts): merge flush + read into artifacts.ts
Three files (artifacts-flush.ts, artifacts-read.ts, artifacts-types.ts)
described one concept — the per-session Artifacts repo. The split between
flush and read was by verb, not by responsibility; each had its own
tokenSecret() helper, and ARTIFACTS_TOKEN_TTL_SECONDS already carried a
'single source of truth' comment that signalled the boundary was wobbly.

Merge artifacts-flush.ts + artifacts-read.ts into src/artifacts.ts.
Single-source tokenSecret(). Keep artifacts-types.ts separate — it's
a stub for @cloudflare/workers-types beta types that will eventually
ship upstream; the file's existence is justified BY its temporariness.

Per the improve-codebase-architecture audit, candidate #5.
